+++ src/app/tests/TestAttributeValueEncoder.cpp
@@ -98,7 +98,7 @@ void TestEncodeListOfBools2(nlTestSuite * aSuite, void * aContext)
 {
     TestSetup test(aSuite);
     bool list[]    = { true, false };
-    CHIP_ERROR err = test.encoder.EncodeList([&list](const TagBoundTLVEncoder & encoder) -> CHIP_ERROR {
+    CHIP_ERROR err = test.encoder.EncodeList([&list](const TagBoundEncoder & encoder) -> CHIP_ERROR {
         for (auto & item : list)
         {
             ReturnErrorOnFailure(encoder.Encode(item));
